requirements.txt
fastapi
uvicorn[standard]
langchain==0.3.25
langchain-openai
langgraph==0.3.34
langgraph-swarm==0.0.10
Markdown
python-dotenv
Requests
supabase
typing
Jinja2
python-multipart
langchain-core
